mcp/requirements.txt
# MCP + RAG System Requirements

# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0

# Database & Storage
supabase==2.0.3
psycopg2-binary==2.9.9
pgvector==0.2.3

# LLM & Embeddings
openai==1.3.5
anthropic==0.7.0
tiktoken==0.5.1

# Vector Operations
numpy>=1.26.0,<2.0
scikit-learn>=1.5.0,<1.6

# Data Processing
pandas==2.1.3
requests==2.31.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3

# Logging & Monitoring
structlog==23.2.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.1
